Reorganise classes into neater packages
