+++ src/main/java/duke/command/TodoCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 import duke.task.Task;
 import duke.task.ToDo;
 import java.io.IOException;
